[ECO-643] Opentok deprecated methods and general cleanup

+++ ib/src/main/java/com/tokbox/android/IB/events/EventUtils.java
@@ -48,7 +48,7 @@ public class EventUtils {
     }
 
     private static int textWidth(String str) {
-        return (int) (str.length() - str.replaceAll(NON_THIN, "").length() / 2);
+        return str.length() - str.replaceAll(NON_THIN, "").length() / 2;
     }
 
     public static String ellipsize(String text, int max) {
